# web/libs/data-access/src/lib/api/models/checkout-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CheckoutLine } from '../models/checkout-line';
export interface CheckoutRequest {
  customerId: string;
  deliveryPostcode: string;
  lines: Array<CheckoutLine>;
  sellerId: string;
}