# web/libs/data-access/src/lib/api/models/order-line-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface OrderLineDto {
  lineTotal: number;
  name: string;
  productId: string;
  quantity: number;
  sku: string;
  unitPrice: number;
}