# web/libs/data-access/src/lib/api/models/product-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CategoryDto } from '../models/category-dto';
export interface ProductDto {
  category: CategoryDto;
  createdAtUtc: string;
  currency: string;
  description: string | null;
  id: string;
  modifiedAtUtc: string | null;
  name: string;
  onHandQuantity: number;
  price: number;
  reservedQuantity: number;
  sellerId: string;
  sku: string;
  slug: string;
  status: string;
}